requirements.txt
deta
fastapi
numpy
python-multipart
jinja2
scipy
requests
uvicorn
aiofiles
pytest
openai
python-dotenv
markdown
html_sanitizer